refactor(agent): move agent build logic out of setters
The `build` method was removed from the builder setters module, leaving only the setter implementations there. The construction logic now lives in its own module so the setters file stays focused on configuration.

Auto-committed-on: macbook
